Added hashing function and deprecated PGCRYPTO.
