+++ extension/src/model/mod.rs
@@ -1,3 +1,4 @@
 pub mod source_objects;
 pub mod dv_schema;
 pub mod queries;
+pub mod prompt_template;
